Move timer to separated file
